src/requirements.txt
Flask
Flask-Cors
Flask-RESTful
Flask-SQLAlchemy
Flask-Session
Flask-Dance
oauthlib
python-dotenv
markdown
matplotlib
pygal